objtool: check: find pc-relative switch jump tables
Instead of having a jump table which stores the actual targets, only
offsets to the current location might be stored in a table and a
relative jump with the according value of that table performed. GCC
for example uses this for position independant binaries. This commit
provides support for objtool of such offset based switch jump tables.